requirements.txt
# Core Dependencies
openai>=1.0.0
anthropic>=0.18.0
requests>=2.31.0
aiohttp>=3.9.0
rich>=13.7.0
click>=8.1.7
python-dotenv>=1.0.0

# Multimodal Processing
pillow>=10.0.0
opencv-python>=4.8.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0

# Security Testing
garak>=0.9.0

# Payload Tools
pwntools>=4.12.0

# Advanced Adversarial Tools
foolbox>=3.3.0
